backend/app/pyproject.toml
[tool.poetry]
name = "facebook-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
pillow = "^10.4.0"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
requests = {extras = ["socks"], version = "^2.32.3"}
faker = "^26.0.0"
fake-useragent = "^1.5.1"
playwright = "^1.45.1"
pytest-playwright = "^0.5.1"
stem = "^1.8.2"
pysocks = "^1.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
